Move ConsistencyCheckJob legacy code to under namespace Old
To be able to keep this working code and be able to come up with
a better solution and have both running in parallel, we moved it to
under the `Old` namespace keeping the old interface via rake tasks
untouched
